Added view for importing schools
Importing schools takes a while (around 10 minutes) so using a console
command on a deployed instance gets timed out. Therefore, a command to
hit the URL can be added to the task queue, which times out after 10
minutes. Documentation found here:
https://developers.google.com/appengine/docs/python/taskqueue/overview-p
ush -- all that's needed in the console is the taskqueue.add() command